Create_Rule
To create the rule: Choose Rules and then provide a name for the rule. I used JavaScript to implement the rule. The code checks to see if the JWT token being requested is for a specific API and then depending on the Identity provider used for authentication, adds a Scope to the access token.