Create Amazon DynamoDB row-level authorization
The following code snippet updates the status attribute of an item in the preceding DynamoDB table. I pass the JWT Id token provided from the user pool in exchange for IAM credentials. The following code assumes that a user is already signed in to your app. I use the Amazon Cognito APIs to get information about the signed in user. For example, by calling var cognitoUser = userPool.getCurrentUser(); in the following code sample, we get the current signed in user. By calling session.getIdToken().getJwtToken() we get the JWT Id token.